+++ src/query/usePartyMutation.tsx
@@ -1,5 +1,4 @@
-import { addMember, addRecruit } from '@/api/party-api';
-import { PartyMember } from '@/types/parties.types';
+import { addRecruit } from '@/api/party-api';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 // 구인글
@@ -11,20 +10,6 @@ export const useAddRecruitMutation = () => {
     onSuccess: () => {
       queryClient.invalidateQueries({
         queryKey: ['recruits'],
-      });
-    },
-  });
-};
-
-// 파티멤버
-export const useAddMemberMutation = (id: PartyMember['party_id']) => {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: addMember,
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: ['members', id],
       });
     },
   });
