+++ src/query/usePartyQuery.ts
@@ -1,5 +1,4 @@
-import { getMembers, getRecruits } from '@/api/party-api';
-import { PartyMember } from '@/types/parties.types';
+import { getRecruits } from '@/api/party-api';
 import { createClient } from '@/utils/supabase/client';
 import { useQuery } from '@tanstack/react-query';
 
@@ -9,14 +8,5 @@ export const useRecruitQuery = () => {
   return useQuery({
     queryKey: ['recruits'],
     queryFn: () => getRecruits(browserClient),
-  });
-};
-
-export const useMemberQuery = (id: PartyMember['party_id']) => {
-  const browserClient = createClient();
-
-  return useQuery({
-    queryKey: ['members', id],
-    queryFn: () => getMembers(browserClient, id),
   });
 };
